requirements.txt
kafka-python==2.0.2
psycopg2==2.9.5
distlib==0.3.6
filelock==3.8.0
platformdirs==2.5.2
psycopg2-binary==2.9.4
virtualenv==20.16.5
requests~=2.28.1